release: v3.47.1 — manual-update ZIP, монтаж для Отлива/Козырька, дробное количество
- create-manual-update.ps1: portable ZIP для ручного обновления
  с README_ОБНОВЛЕНИЕ.txt, не требует GitHub Releases
- Отлив и Козырёк: монтаж по периметру (500/750 ₽/м.п.), по умолчанию выключен
- Дробное количество: OrderItem.Quantity изменён с int на double,
  добавлен QuantityConverter
- UI: прозрачный фон SignToggleCheckBox без цветного фона
- update-log: 3.47.1 "Технические исправления"
- docs: CURRENT_STATE.md, AUTO_UPDATE.md, RELEASE_PROCESS.md синхронизированы

+++ MosquitoNetCalculator/Models/OrderItem.Dto.cs
@@ -15,7 +15,7 @@ namespace MosquitoNetCalculator.Models
         public string Color { get; set; } = "";
         public double Width { get; set; }
         public double Height { get; set; }
-        public int Quantity { get; set; } = 1;
+        public double Quantity { get; set; } = 1;
         public double Price { get; set; }
         public int InstallationMode { get; set; }
         public bool HasInstallation { get; set; } = true;
